genai_microservice_rag/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.5
faiss-cpu>=1.8.0
openai>=1.35.0
tiktoken>=0.7.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
redis>=5.0.4
prometheus-client>=0.20.0
aiofiles>=23.2.1
httpx>=0.27.0
pytest>=7.4.0
